requirements.txt
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
httpx>=0.27.0
rank-bm25>=0.2.2
pymupdf>=1.24.0
pdfplumber>=0.11.0
pytesseract>=0.3.10
Pillow>=10.0.0
pyarrow>=14.0.0
matplotlib>=3.8.0

# ColPali / ColQwen 本地部署
torch>=2.1.0
transformers>=4.45.0
colpali-engine @ git+https://github.com/illuin-tech/colpali.git

# Milvus Lite
pymilvus>=2.4.0
setuptools>=70.0.0

# 本地 BGE-M3 向量
FlagEmbedding>=1.3.5

